Modify JWT generators to accept custom claims
Before this change, GeneratorInterface->getToken() only accepts a
subject. This is inflexible with respect to being able to store
custom claims in generated tokens. It also goes against the JWT
spec, which states that the subject is an optional claim.

This change modifies getToken() to take an optional associative
array of claims and internally adds a common set of claims to it
where they are not already set.